apps/recommendations/requirements.txt
flask
psycopg2-binary
python-dotenv
pika
transformers
torch
beautifulsoup4
flask_cors
alt-profanity-check
sqlalchemy